Added Tox Unit Tests to Check For Content Type Key in Exclude Dictionary
Added LmsSearchFilterGenerator mock class to avoid creating a django app. Creating a django app was resulting in many dependencies / django apps. In addition, it was resulting in an error. It was more convenient to unit test the ExcludeCapaSearchFilterGenerator with a mock LmsSearchFilterGenerator mock class.

The LmsSearchFilterGenerator and the SearchFilterGenerator already have unit tests, thus, the unit test only for ExcludeCapaSearchFilterGenerator.