SE-2672 Added Tox Tests and Updated to Version 0.2 and Removed Python 2.7 Support (#1)
Added tox tests and updated the extension to python 3, removing python 2.7 support.

* Added Tox Unit Tests to Check For Content Type Key in Exclude Dictionary

Added LmsSearchFilterGenerator mock class to avoid creating a django app. Creating a django app was resulting in many dependencies / django apps. In addition, it was resulting in an error. It was more convenient to unit test the ExcludeCapaSearchFilterGenerator with a mock LmsSearchFilterGenerator mock class.

The LmsSearchFilterGenerator and the SearchFilterGenerator already have unit tests, thus, the unit test only for ExcludeCapaSearchFilterGenerator.

* Added PylintRc from EdxLint and Updated Files to Fix Pylint Warnings

* Updated Setup to Version 0.2 and Removed Python 2.7 Support

* Updated README to Add Testing Instruction

* Updated Tox File to Use Juniper's Django Version

Co-authored-by: Jillian Vogel <jill@opencraft.com>

* Changed Content Type Assertion to Check For CAPA Content Type

Co-authored-by: Jillian Vogel <jill@opencraft.com>

* Used Default Settings for Pylint with Py38-Quality Test

Had to disable 'too-few-public-methods- warning for ExcludeCapaSearchFilterGenerator

* Added Mock LmsSearchFilterGenerator After Updating GitIgnore

GitIgnore was ignoring all lib/ paths; therefore, added an exception for tests/utils/lms/lib

* Updated Super Call for Exclude Dictionary

Ended up removing the SearchFilterGenerator as it was no longer needed.

Co-authored-by: Jillian Vogel <jill@opencraft.com>